code cov fix

+++ localfs/store.go
@@ -486,5 +486,14 @@ func (o *object) Close() error {
 }
 
 func (o *object) Release() error {
-	return os.Remove(o.cachepath)
+	if o.cachedcopy != nil {
+		gou.Debugf("release %q vs %q", o.cachedcopy.Name(), o.cachepath)
+		o.cachedcopy.Close()
+		o.cachedcopy = nil
+		o.opened = false
+		return os.Remove(o.cachepath)
+	}
+	// most likely this doesn't exist so don't return error
+	os.Remove(o.cachepath)
+	return nil
 }
